requirements.txt
greenlet==3.1.1
perlin-noise==1.13
psycopg2-binary==2.9.9
python-dotenv==1.0.1
SQLAlchemy==2.0.35
typing_extensions==4.12.2
